Refactor some html to be semantic, and some additional organization (excluding <main/>)
