+++ src/actions/auth.js
@@ -1,16 +1,8 @@
 import jwtDecode from 'jwt-decode';
 import {SubmissionError} from 'redux-form';
-import { API_BASE_URL } from '../config';
-import {normalizeResponseErrors} from './utils';
 
-// Puts full authToken into Redux state
-// export const SET_TOKEN = 'SET_TOKEN';
-// export const setToken = token => {
-// 	return {
-// 		type: SET_TOKEN,
-// 		token
-// 	};
-// };
+import {API_BASE_URL} from '../config';
+import {normalizeResponseErrors} from './utils';
 
 // Removes authToken & user info from Redux state
 export const CLEAR_TOKEN = 'CLEAR_TOKEN';
@@ -50,7 +42,6 @@ export const loginError = error => {
 //Store in localStorage & decompose into state
 export const storeToken = (token, dispatch) => {
 	const decodedToken = jwtDecode(token);
-	// dispatch(setToken(decodedToken));
 	// set local storage BEFORE sending decoded token to avoid timing errors
 	localStorage.setItem('authToken', token);
 	dispatch(loginSuccess(decodedToken.user));
