+++ src/components/Forms/field.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+
 import './styles/forms.css';
 
 export const Field = props => {
 	const Element = props.element || 'input';
 	return (
-		<div className="form-field" id={props.id}>
+		<fieldset className="form-field" id={props.id}>
 			<label htmlFor={props.input.name}>{props.label}</label>
 			<Element
 				className={(props.meta.error && props.meta.touched) ? 'error-border' : 'placeholder-class'}
@@ -13,19 +14,12 @@ export const Field = props => {
 				value={props.value}
 				aria-required="true"
 				placeholder={props.placeholder ? props.placeholder : null }
-				// defaultValue={props.defaultValue ? props.defaultValue : null}
 				{...props.input}
 			>
 				{props.children}
 			</Element>
-			{/*
-        If error evaluates to true -> show span element
-        if error is undefined -> React ignores the span element
-      */}
-			{props.meta.error &&
-      props.meta.touched &&
-      <div className="form-field-error" aria-live="polite" role="alert">{props.meta.error}</div>}
-		</div>
+			{props.meta.error && props.meta.touched && <div className="form-field-error" aria-live="polite" role="alert">{props.meta.error}</div>}
+		</fieldset>
 	);
 };
