+++ src/components/forms/styles/forms.css
@@ -7,9 +7,6 @@
     }
 }
 
-
-
-
 .form-wrapper {
     margin: 0 auto;
     padding: 20px;
