+++ src/components/schedule-view.js
@@ -1,8 +1,9 @@
 import React from 'react';
-import { connect } from 'react-redux';
+import {connect} from 'react-redux';
 import moment from 'moment';
-import {getThisMonth, getThisWeek, getToday} from '../actions/utils';
-import { setFramesView, fetchFrames } from '../actions/frames';
+
+import {getThisMonth,getThisWeek,getToday} from '../actions/utils';
+import {setFramesView,fetchFrames} from '../actions/frames';
 
 import './styles/schedule-view.css';
 
@@ -41,7 +42,7 @@ export class ScheduleView extends React.Component {
 		return (
 			<div className="view-filter">
 				<h2>{startSchedule} - {endSchedule}</h2>
-				<div className="red-border"> 
+				<div className="red-border">
 					<button className={this.props.view === 'daily' ? 'active view-btn' : 'view-btn'} title="Daily Schedule"
 						onClick={() => this.handleDaily()}
 					>
