+++ src/components/styles/app.css
@@ -104,7 +104,6 @@
   transform: rotate(180deg);
 }
 
-
 @keyframes App-logo-spin {
   from { transform: rotate(0deg); }
   to { transform: rotate(360deg); }
