+++ src/components/styles/employees.css
@@ -1,7 +1,6 @@
 .employees-wrapper {
     position: relative;
     top: 50px;
-    /*height: 130vh;*/
     color: #edeef0;
     background-color: #2b2d3d;
     padding: 5%;
