+++ src/components/styles/landing-page.css
@@ -3,7 +3,6 @@
 	margin: 0 auto;
 	position: relative;
 	top: 50px;
-
 	/*background-color: #FFFFFF;*/
 	/*text-align: center;*/
 	/*color: #edeef0;*/
