+++ src/components/styles/profile.css
@@ -10,7 +10,6 @@
 .profile-wrapper {
     position: relative;
     top: 50px;
-    /*height: 130vh;*/
     color: #edeef0;
     background-color: #2b2d3d;
     padding: 5%;
